Replace vim-autocomplpop with supertab. Change supertab configs:
Note: a special value of 'context' is supported which will result in
super tab attempting to use the text preceding the cursor to decide which
type of completion to attempt.  Currently super tab can recognize method
calls or attribute references via '.', '::' or '->', and file path
references containing '/'.

let g:SuperTabDefaultCompletionType = "context"

Why I changed?
Auto pop up is bit of annoying and without change the configuration of
vim-autocomplpop I trys supertab. It does not conflict with snippetMate,
which is the initial reason that I did not use supertab before. In
addtion, I found the context option which make supertab work the same
with the effect vim-autocomplpop provids me, but not annoying.
So I changed. Maybe some other day I will take a close look at
vim-autocomplpop.